requirements.txt
albumentations==1.3.1
certifi==2023.11.17
charset-normalizer==3.3.2
cycler==0.11.0
fonttools==4.38.0
idna==3.6
imageio==2.31.2
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.7.1
urllib3==2.0.7
